Move autoscaling e2e tests to a separate directory
