+++ test/e2e/BUILD
@@ -14,7 +14,6 @@ go_library(
         "addon_update.go",
         "apparmor.go",
         "cadvisor.go",
-        "cluster_size_autoscaling.go",
         "cluster_upgrade.go",
         "cronjob.go",
         "daemon_restart.go",
@@ -23,7 +22,6 @@ go_library(
         "deployment.go",
         "disruption.go",
         "dns.go",
-        "dns_autoscaling.go",
         "dns_common.go",
         "dns_configmap.go",
         "e2e.go",
@@ -39,7 +37,6 @@ go_library(
         "gke_local_ssd.go",
         "gke_node_pools.go",
         "ha_master.go",
-        "horizontal_pod_autoscaling.go",
         "ingress.go",
         "initial_resources.go",
         "job.go",
@@ -250,6 +247,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
+        "//test/e2e/autoscaling:all-srcs",
         "//test/e2e/chaosmonkey:all-srcs",
         "//test/e2e/cluster-logging:all-srcs",
         "//test/e2e/common:all-srcs",
