+++ test/e2e/autoscaling/OWNERS
@@ -1,0 +1,8 @@
+reviewers:
+  - mwielgus
+  - jszczepkowski
+  - MaciekPytel
+approvers:
+  - mwielgus
+  - jszczepkowski
+  - MaciekPytel
