+++ test/e2e/autoscaling/horizontal_pod_autoscaling.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e
+package autoscaling
 
 import (
 	"time"
