+++ test/e2e/e2e_test.go
@@ -19,6 +19,7 @@ package e2e
 import (
 	"testing"
 
+	_ "k8s.io/kubernetes/test/e2e/autoscaling"
 	_ "k8s.io/kubernetes/test/e2e/cluster-logging"
 	"k8s.io/kubernetes/test/e2e/framework"
 	_ "k8s.io/kubernetes/test/e2e/perf"
